Integrated the test scripts in
