--- Assets/Scripts/Combat/PlayerTargetScanner.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 15b34655bffe8e14195a89f563a47e87
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
