+++ Assets/Scripts/Combat/Weapons/MeleeWeapon.cs
@@ -1,4 +1,5 @@
 using System;
+using ScalePact.Core;
 using ScalePact.Utils;
 using Unity.VisualScripting;
 using UnityEngine;
