--- Assets/Scripts/Enemies/EnemyBehaviorStates.cs
@@ -1,15 +1,0 @@
-namespace ScalePact.Enemies
-{
-    public enum EnemyBehaviorState
-    {
-        NoState,
-        PatrolState,
-        SuspicionState,
-        ChaseState,
-        AttackState,
-        AttackIdleState,
-        ImpactState,
-        DeathState,
-
-    }
-}
