--- Assets/Scripts/Enemies/EnemyBehaviorStates.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 48de5e73d3d5f524b8137e7a357c93bd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
