--- Assets/Scripts/Enemies/EnemyCombat.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: daff2a4e7ae1e914180ee93cca7d2273
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
