+++ Assets/Scripts/Enemies/SimpleEnemyStates/SimpleEnemyDeathState.cs
