+++ Assets/Scripts/Enemies/SimpleEnemyStates/SimpleEnemyIdleState.cs.meta
