+++ Assets/Scripts/Player/PlayerCombat.cs
@@ -1,4 +1,7 @@
 using System.Collections;
+using System.Collections.Generic;
+using System.Linq;
+using Cinemachine;
 using ScalePact.Combat;
 using ScalePact.Core;
 using ScalePact.Core.Input;
@@ -10,28 +13,41 @@ namespace ScalePact.Player
 {
     public class PlayerCombat : MonoBehaviour
     {
+        [Header("Targeting Variables")]
+        [SerializeField] float targettingRadius;
+        [SerializeField] LayerMask targettingLayer;
+
+        [Header("Attacks and Abilities")]
         [SerializeField] MeleeWeapon[] meleeWeaponCombo;
 
         public bool IsAttacking { get; private set; } = false;
+        public Collider ActiveTarget { get; private set; }
+        public List<Collider> TargetColliders { get; private set; } = new();
 
+        int targetIndex;
+        bool isLockedOn;
         float comboMinDelay = 0.1f;
         int comboMaxIndex = 2;
         int comboIndex;
-        Coroutine comboAttackResetCoroutine;
 
+        Collider[] overlappedColliders;
+
+        GameObject targetFollow;
+        CinemachineTargetGroup targetGroup;
+
+        Coroutine comboAttackResetCoroutine;
         InputManager inputManager;
         Animator animator;
-        PlayerTargetScanner targetScanner;
         PlayerForceReceiver forceReceiver;
-        Health health;
+        Damageable damageable;
 
         private void Awake()
         {
             inputManager = GetComponent<InputManager>();
             animator = GetComponent<Animator>();
-            targetScanner = GetComponent<PlayerTargetScanner>();
             forceReceiver = GetComponent<PlayerForceReceiver>();
-            health = GetComponent<Health>();
+            damageable = GetComponentInChildren<Damageable>();
+            targetGroup = FindObjectOfType<CinemachineStateDrivenCamera>().GetComponentInChildren<CinemachineTargetGroup>();
 
             comboIndex = -1;
             comboAttackResetCoroutine = null;
@@ -40,25 +56,45 @@ namespace ScalePact.Player
         private void OnEnable()
         {
             inputManager.LightAttackEvent += OnLightAttackPressed;
+            inputManager.ToggleTargetEvent += OnLockOnTarget;
+            inputManager.SwitchTargetEvent += OnSwitchTarget;
         }
 
         private void OnDisable()
         {
             inputManager.LightAttackEvent -= OnLightAttackPressed;
+            inputManager.ToggleTargetEvent -= OnLockOnTarget;
+            inputManager.SwitchTargetEvent -= OnSwitchTarget;
+        }
+
+        private void Start()
+        {
+            targetFollow = new()
+            {
+                name = "TargetFollow"
+            };
+
+            targetGroup.AddMember(targetFollow.transform, 1, 1);
         }
 
         private void Update()
         {
+            GetPotentialTargets();
 
+            if (isLockedOn)
+            {
+                targetFollow.transform.position = Vector3.Lerp(targetFollow.transform.position, ActiveTarget.bounds.center, Time.deltaTime * 4f); //HISS magic number
+            }
         }
 
+        #region Attacks
         private void OnLightAttackPressed()
         {
-            if (health.IsDead) return;
+            if (damageable.IsDead) return;
 
             IsAttacking = true;
 
-            Collider closestTarget = targetScanner.GetClosestTargetNoTargetting();
+            Collider closestTarget = GetClosestTargetNoTargetting();
             if (closestTarget != null)
             {
                 forceReceiver.FaceTarget(closestTarget.transform);
@@ -103,8 +139,102 @@ namespace ScalePact.Player
             //forceReceiver.AddForce(transform.forward * attackData[index].AttackForce);
             animator.SetInteger(PlayerHashIDs.AttackIndexHash, index);
         }
+        #endregion
 
-        //Anim Events
+        #region Abilities
+        //ABILITES HERE?
+        #endregion
+
+        #region Target Scanner
+        public Collider GetClosestTargetNoTargetting()
+        {
+            return GetClosestTarget();
+        }
+
+        private void GetPotentialTargets()
+        {
+            overlappedColliders = Physics.OverlapSphere(
+                transform.position, targettingRadius, 
+                targettingLayer, QueryTriggerInteraction.Ignore);
+
+            TargetColliders.Clear();
+
+            for (int i = 0; i < overlappedColliders.Length; i++)
+            {
+                var overlapped = 
+                    overlappedColliders[i].GetComponentInChildren<Damageable>();
+                if (overlapped != null)
+                {
+                    TargetColliders.Add(overlappedColliders[i]);
+                }
+            }
+        }
+
+        void SortTargetList()
+        {
+            TargetColliders = TargetColliders.OrderBy(
+                x => Camera.main.WorldToScreenPoint(x.transform.position).x).ToList();
+        }
+
+        Collider GetClosestTarget()
+        {
+            if (TargetColliders.Count == 0) return null;
+
+            Collider bestTarget = null;
+            TargetColliders = TargetColliders.OrderBy(
+                x => Vector3.Distance(transform.position, x.transform.position)).ToList();
+            if (TargetColliders[0] != null)
+            {
+                bestTarget = TargetColliders[0];
+            }
+            return bestTarget;
+        }
+
+        bool CheckForTargets()
+        {
+            if (TargetColliders.Count > 0)
+            {
+                return true;
+            }
+            return false;
+        }
+        #endregion
+
+        #region Event Callbacks
+        void OnLockOnTarget()
+        {
+            if (!CheckForTargets())
+            {
+                isLockedOn = false;
+                return;
+            }
+
+            isLockedOn = !isLockedOn;
+            if (isLockedOn)
+            {
+                ActiveTarget = GetClosestTarget();
+            }
+        }
+
+        void OnSwitchTarget()
+        {
+            if (isLockedOn)
+            {
+                SortTargetList();
+                if (targetIndex < (TargetColliders.Count - 1))
+                {
+                    targetIndex = targetIndex + 1;
+                }
+                else
+                {
+                    targetIndex = TargetColliders.Count - 1;
+                }
+                ActiveTarget = TargetColliders[targetIndex];
+            }
+        }
+        #endregion
+
+        #region Anim Events
         void EnableCollider()
         {
             // if (comboIndex == -1) return;
@@ -132,5 +262,14 @@ namespace ScalePact.Player
                 item.EndAttack();
             }
         }
+        #endregion
+
+#if UNITY_EDITOR
+        private void OnDrawGizmos()
+        {
+            Gizmos.color = Color.blue;
+            Gizmos.DrawWireSphere(transform.position, targettingRadius);
+        }
+#endif
     }
 }
