--- Assets/_TESTS/EnemyMovementNEW.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 98290a97557daef41a34389ba606856b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
